+++ HISTORY.md
@@ -3,6 +3,7 @@
 ## Done
 - [x] rewrite test server in c, test cases in golang
 - [x] update travis ci with new test env
+- [x] implement timeout.set, timeout.get functions
 
 ---